refactor(lint)!: remove all #[allow] attributes, fix root causes (part 1)
Cargo.toml [lints] exceptions were removed by hand; this starts fixing every
lint at the source instead of suppressing it:

- casts: min_runtime_ms config type u32 → f32 (matches exec_time/FilterBuilder;
  serde rejects negative/non-finite), offsets i64↔u64 via try_from, timestamp
  days via try_from, rec/s via u32::try_from + f64::from, filtered-percentage
  via integer permille arithmetic
- unused_async: the stats chain (scan_files → run_stats → handle_stats) had no
  await points — made fully synchronous; scan_files' always-Ok Result dropped
- dead_code: truncate_to_120_chars, classify_error_kind,
  add_parse_error_with_kind deleted (zero callers) along with their tests
- items_after_statements: consts/types/uses hoisted to scope tops; SimpleLogger
  moved out of init_logging to module level
- similar_names: idx_u64/idx_i64 → count_seed/ms_seed

Remaining (next commits): missing_errors_doc, too_many_arguments,
type_complexity, bin/lib duplicate-compile unused_imports,
multiple_crate_versions.

+++ src/engine/prepare.rs
@@ -106,8 +106,7 @@ pub(super) fn make_progress_bar(show_progress: bool, total_files: usize) -> Opti
 pub(super) fn build_indicator_filters(indicators: &IndicatorFilters) -> Vec<Filter> {
     let mut filters = Vec::new();
     if let Some(min_ms) = indicators.min_runtime_ms {
-        #[allow(clippy::cast_precision_loss)]
-        filters.push(FilterBuilder::new().exec_time_gte(min_ms as f32).build());
+        filters.push(FilterBuilder::new().exec_time_gte(min_ms).build());
     }
     if let Some(min_r) = indicators.min_row_count {
         // rowcount >= min_r: for u32, rowcount_gt(min_r - 1) works when min_r > 0
